extract shared component to sign transactions
