+++ deployables/app/app/routes/sign/account.$accountId.$transactions/sign.tsx
@@ -1,9 +1,8 @@
 import { authorizedAction, authorizedLoader } from '@/auth'
 import { ConnectWallet } from '@/components'
 import { simulateTransactionBundle } from '@/simulation-server'
-import { jsonRpcProvider, parseTransactionData, routeTitle } from '@/utils'
+import { parseTransactionData, routeTitle } from '@/utils'
 import { invariantResponse } from '@epic-web/invariant'
-import { EXPLORER_URL } from '@zodiac/chains'
 import {
   dbClient,
   getAccount,
@@ -11,43 +10,17 @@ import {
   toExecutionRoute,
 } from '@zodiac/db'
 import { getBoolean, getNumberMap } from '@zodiac/form-data'
-import { useIsPending } from '@zodiac/hooks'
-import {
-  CompanionAppMessageType,
-  type CompanionAppMessage,
-} from '@zodiac/messages'
 import { checkPermissions, isValidRoute, queryRoutes } from '@zodiac/modules'
 import { isUUID } from '@zodiac/schema'
-import {
-  Error,
-  errorToast,
-  Form,
-  PrimaryButton,
-  Success,
-  successToast,
-  Warning,
-} from '@zodiac/ui'
-import type { Eip1193Provider } from 'ethers'
-import {
-  ArrowDownToLine,
-  ArrowLeftRight,
-  ArrowUpFromLine,
-  SquareArrowOutUpRight,
-} from 'lucide-react'
-import { Suspense, useEffect } from 'react'
-import { Await, useActionData, useLoaderData } from 'react-router'
-import {
-  execute,
-  ExecutionActionType,
-  planExecution,
-  prefixAddress,
-  unprefixAddress,
-  type ExecutionState,
-} from 'ser-kit'
-import { useAccount, useConnectorClient } from 'wagmi'
+import { Error, Form, Success, Warning } from '@zodiac/ui'
+import { ArrowDownToLine, ArrowLeftRight, ArrowUpFromLine } from 'lucide-react'
+import { Suspense } from 'react'
+import { Await } from 'react-router'
+import { planExecution, prefixAddress } from 'ser-kit'
 import { getDefaultNonces } from '../getDefaultNonces'
 import { revokeApprovalIfNeeded } from '../revokeApprovalIfNeeded'
 import { ApprovalOverviewSection, ReviewAccountSection } from '../sections'
+import { SignTransaction } from '../SignTransaction'
 import { SkeletonFlowTable, TokenTransferTable } from '../table'
 import type { Route } from './+types/sign'
 
@@ -105,13 +78,11 @@ export const loader = async (args: Route.LoaderArgs) =>
         isValidRoute: isValidRoute(queryRoutesResult),
         hasQueryRoutesError: queryRoutesResult.error != null,
         id: route.id,
-        initiator: route.wallet.address,
-        avatar: executionRoute.avatar,
-        chainId: account.chainId,
+        account,
+        wallet: route.wallet,
         simulation: simulate(),
         permissionCheck: permissionCheckResult.permissionCheck,
         waypoints: route.waypoints,
-        metaTransactions,
         defaultSafeNonces: getDefaultNonces(plan),
       }
     },
@@ -191,10 +162,9 @@ export const action = async (args: Route.ActionArgs) =>
 
 const SubmitPage = ({
   loaderData: {
-    initiator,
-    chainId,
+    wallet,
     id,
-    avatar,
+    account,
     waypoints,
     isValidRoute,
     permissionCheck,
@@ -202,6 +172,7 @@ const SubmitPage = ({
     hasQueryRoutesError,
     defaultSafeNonces,
   },
+  actionData,
 }: Route.ComponentProps) => {
   return (
     <Form>
@@ -219,7 +190,7 @@ const SubmitPage = ({
                   <TokenTransferTable
                     title="Tokens Sent"
                     columnTitle="To"
-                    avatar={avatar}
+                    avatar={account.address}
                     icon={ArrowUpFromLine}
                     tokens={sent}
                   />
@@ -227,7 +198,7 @@ const SubmitPage = ({
                   <TokenTransferTable
                     title="Tokens Received"
                     columnTitle="From"
-                    avatar={avatar}
+                    avatar={account.address}
                     icon={ArrowDownToLine}
                     tokens={received}
                   />
@@ -235,7 +206,7 @@ const SubmitPage = ({
                   <TokenTransferTable
                     title="Other Token Movements"
                     columnTitle="From → To"
-                    avatar={avatar}
+                    avatar={account.address}
                     icon={ArrowLeftRight}
                     tokens={other}
                   />
@@ -284,7 +255,7 @@ const SubmitPage = ({
           id={id}
           isValidRoute={isValidRoute}
           hasQueryRoutesError={hasQueryRoutesError}
-          chainId={chainId}
+          chainId={account.chainId}
           waypoints={waypoints}
           defaultSafeNonces={defaultSafeNonces}
         />
@@ -294,145 +265,22 @@ const SubmitPage = ({
         title="Pilot Signer"
         description="Make sure that your wallet is connected to the route's operator account."
       >
-        <ConnectWallet chainId={chainId} pilotAddress={initiator} />
+        <ConnectWallet
+          chainId={account.chainId}
+          pilotAddress={wallet.address}
+        />
       </Form.Section>
 
       <Form.Actions>
-        <SubmitTransaction />
+        <SignTransaction
+          chainId={account.chainId}
+          walletAddress={wallet.address}
+          safeAddress={prefixAddress(account.chainId, account.address)}
+          executionPlan={actionData == null ? null : actionData.plan}
+        />
       </Form.Actions>
     </Form>
   )
 }
 
 export default SubmitPage
-
-type SubmitTransactionProps = {
-  disabled?: boolean
-}
-
-const SubmitTransaction = ({ disabled = false }: SubmitTransactionProps) => {
-  const { chainId, avatar, initiator } = useLoaderData<typeof loader>()
-  const walletAccount = useAccount()
-  const { data: connectorClient } = useConnectorClient()
-
-  const actionData = useActionData<typeof action>()
-
-  useEffect(() => {
-    if (actionData == null) {
-      return
-    }
-
-    const { plan } = actionData
-
-    const executePlan = async () => {
-      const state: ExecutionState = []
-      try {
-        await execute(plan, state, connectorClient as Eip1193Provider, {
-          origin: 'Zodiac Pilot',
-        })
-
-        const safeTxHash =
-          state[
-            plan.findIndex(
-              (action) =>
-                action.type === ExecutionActionType.PROPOSE_TRANSACTION,
-            )
-          ]
-
-        if (safeTxHash != null) {
-          console.debug(
-            `Transaction batch has been proposed with safeTxHash ${safeTxHash}`,
-          )
-
-          const url = new URL('/transactions/tx', 'https://app.safe.global')
-
-          url.searchParams.set('safe', avatar)
-          url.searchParams.set(
-            'id',
-            `multisig_${unprefixAddress(avatar)}_${safeTxHash}`,
-          )
-
-          successToast({
-            title: 'Transaction batch has been proposed for execution',
-            message: (
-              <a
-                href={url.toString()}
-                className="inline-flex items-center gap-1"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <SquareArrowOutUpRight size={16} />
-                {'View in Safe{Wallet}'}
-              </a>
-            ),
-          })
-        } else {
-          const txHash =
-            state[
-              plan.findLastIndex(
-                (action) =>
-                  action.type === ExecutionActionType.EXECUTE_TRANSACTION,
-              )
-            ]
-
-          if (txHash) {
-            console.debug(
-              `Transaction batch has been submitted with transaction hash ${txHash}`,
-            )
-            const receipt =
-              await jsonRpcProvider(chainId).waitForTransaction(txHash)
-
-            console.debug(`Transaction ${txHash} has been executed`, receipt)
-
-            successToast({
-              title: 'Transaction batch has been executed',
-              message: (
-                <a
-                  href={new URL(
-                    `tx/${txHash}`,
-                    EXPLORER_URL[chainId],
-                  ).toString()}
-                  className="inline-flex items-center gap-1"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  <SquareArrowOutUpRight size={16} />
-                  View in block explorer
-                </a>
-              ),
-            })
-          }
-        }
-
-        window.postMessage({
-          type: CompanionAppMessageType.SUBMIT_SUCCESS,
-        } satisfies CompanionAppMessage)
-      } catch (error) {
-        console.debug({ error })
-        errorToast({
-          title: 'Error',
-          message: 'Submitting the transaction batch failed',
-        })
-      }
-    }
-
-    executePlan()
-  }, [actionData, avatar, chainId, connectorClient])
-
-  const isSubmitting = useIsPending()
-
-  if (
-    disabled ||
-    walletAccount.chainId !== chainId ||
-    walletAccount.address?.toLowerCase() !== initiator?.toLowerCase() ||
-    connectorClient == null
-  ) {
-    return <PrimaryButton disabled>Sign</PrimaryButton>
-  }
-
-  return (
-    <PrimaryButton submit busy={isSubmitting}>
-      Sign
-    </PrimaryButton>
-  )
-}
